Folded the redundant StripeAssembler into _Strand
