+++ src/cr/cube/stripe/assembler.py
@@ -16,360 +16,7 @@ from cr.cube.collator import (
     SortByValueCollator,
 )
 from cr.cube.enums import COLLATION_METHOD as CM, ORDER_FORMAT
-from cr.cube.stripe.measure import StripeMeasures
 from cr.cube.util import lazyproperty
-
-
-class StripeAssembler:
-    """Provides measures, marginals, and totals for a (1D) strand cube-slice.
-
-    An assembled stripe measure is a 1D ndarray reflecting all ordering, insertion, and
-    hiding transforms applied to the dimension. An assembled margin is often a scalar.
-
-    `cube` is the `cr.cube.Cube` object containing the data for this matrix.
-
-    `rows_dimension` is the Dimension object describing the stripe.
-
-    `slice_idx` is an int offset indicating which portion of `cube` data to use for this
-    stripe.
-    """
-
-    def __init__(self, cube, rows_dimension, ca_as_0th, slice_idx):
-        self._cube = cube
-        self._rows_dimension = rows_dimension
-        self._ca_as_0th = ca_as_0th
-        self._slice_idx = slice_idx
-
-    @lazyproperty
-    def derived_row_idxs(self):
-        """tuple of int index of each derived row-element in this stripe.
-
-        Subtotals cannot be derived
-
-        An element is derived if it's a subvariable of a multiple response dimension,
-        which has been produced by the zz9, and inserted into the response data.
-
-        All other elements, including regular MR and CA subvariables, as well as
-        categories of CAT dimensions, are not derived. Subtotals are also not derived
-        in this sense, because they're not even part of the data (elements).
-        """
-        rows_dim = self._rows_dimension
-        n_subtotals = len(rows_dim.subtotals)
-        derivs = [e.derived for e in rows_dim.valid_elements] + [False] * n_subtotals
-        return tuple(np.where(np.array(derivs)[self._row_order_signed_indexes])[0])
-
-    @lazyproperty
-    def diff_row_idxs(self):
-        """tuple of int index of each difference row-element in this stripe.
-
-        Valid elements are cannot be differences, only some subtotals can.
-        """
-        rows_dim = self._rows_dimension
-        n_valids = len(rows_dim.valid_elements)
-        diffs = [False] * n_valids + [e.is_difference for e in rows_dim.subtotals]
-        return tuple(np.where(np.array(diffs)[self._row_order_signed_indexes])[0])
-
-    @lazyproperty
-    def inserted_row_idxs(self):
-        """tuple of int index of each inserted row in this stripe.
-
-        Provided index values correspond rows after any insertion of subtotals,
-        re-ordering, and hiding/pruning.
-        """
-        return tuple(
-            i for i, row_idx in enumerate(self._row_order_signed_indexes) if row_idx < 0
-        )
-
-    @lazyproperty
-    def means(self):
-        """1D np.float64 ndarray of mean for each row.
-
-        Raises ValueError when the cube-result does not include a means cube-measure.
-        """
-        return self._assemble_vector(self._measures.means.blocks)
-
-    @lazyproperty
-    def payload_order(self):
-        """1D np.int64 ndarray of signed int idx respecting the payload order"""
-        empty_row_idxs = tuple(
-            i for i, N in enumerate(self._measures.pruning_base) if N == 0
-        )
-        return PayloadOrderCollator(self._rows_dimension, empty_row_idxs).payload_order
-
-    @lazyproperty
-    def population_proportions(self):
-        """1D np.float64 population-proportion for each row
-
-        Generally equal to the table_proprotions, but because we don't divide the
-        population when the row is a CAT_DATE, can also be all 1s. Used to calculate
-        the population_counts.
-        """
-        population_proportions = self._assemble_vector(
-            self._measures.population_proportions.blocks
-        )
-        # Diff subtotals not allowed in population measure
-        if self.diff_row_idxs:
-            population_proportions[self.diff_row_idxs] = np.nan
-        return population_proportions
-
-    @lazyproperty
-    def population_proportion_stderrs(self):
-        """1D np.float64 population-proportion-standard-error for each row
-
-        Generally equal to the table_proprotion_standard_error, but because we don't
-        divide the population when the row is a CAT_DATE, can also be all 0s. Used to
-        calculate the population_counts_moe.
-        """
-        return self._assemble_vector(
-            self._measures.population_proportion_stderrs.blocks
-        )
-
-    @lazyproperty
-    def row_count(self):
-        """int count of rows in this stripe.
-
-        This count includes inserted rows but not rows that have been hidden/pruned.
-        """
-        return len(self._row_order_signed_indexes)
-
-    @lazyproperty
-    def row_aliases(self):
-        """1D str ndarray of row alias for each matrix row.
-
-        These are suitable for use as row headings; aliases for subtotal rows appear in
-        the sequence and aliases are ordered to correspond with their respective data
-        row.
-        """
-        return np.array(
-            self._rows_dimension.element_aliases + self._rows_dimension.subtotal_aliases
-        )[self._row_order_signed_indexes]
-
-    @lazyproperty
-    def row_codes(self):
-        """1D int ndarray of row code for each matrix row.
-
-        These are suitable for use as row headings; codes for subtotal rows appear in
-        the sequence and codes are ordered to correspond with their respective data
-        row.
-        """
-        return np.array(
-            self._rows_dimension.element_ids + self._rows_dimension.insertion_ids
-        )[self._row_order_signed_indexes]
-
-    @lazyproperty
-    def row_labels(self):
-        """1D str ndarray of row name for each matrix row.
-
-        These are suitable for use as row headings; labels for subtotal rows appear in
-        the sequence and labels are ordered to correspond with their respective data
-        row.
-        """
-        return np.array(
-            self._rows_dimension.element_labels + self._rows_dimension.subtotal_labels
-        )[self._row_order_signed_indexes]
-
-    def row_order(self, format=ORDER_FORMAT.SIGNED_INDEXES):
-        """1D np.int64 ndarray of idx for each assembled row of stripe.
-
-        If order format is `SIGNED_INDEXES` negative values represent inserted
-        subtotal-row locations; for `BOGUS_IDS` insertios are represented by
-        `ins_{insertion_id}` string.
-        Indices appear in the order rows are to appear in the final result.
-        """
-        # --- specify dtype explicitly to prevent error when display-order is empty. The
-        # --- default dtype is float, which cannot be used to index an array.
-        if format == ORDER_FORMAT.BOGUS_IDS:
-            return self._row_order_bogus_ids
-        return self._row_order_signed_indexes
-
-    @lazyproperty
-    def rows_dimension_fills(self):
-        """tuple of optional RGB str like "#def032" fill color for each strand row.
-
-        Each value reflects the resolved element-fill transform cascade. The length and
-        ordering of the sequence correspond to the rows in the slice, including
-        accounting for insertions, ordering, and hidden rows. A fill value is `None`
-        when no explicit fill color is defined for that row, indicating the default fill
-        color for that row should be used, probably coming from a caller-defined theme.
-        """
-        element_fills = tuple(e.fill for e in self._rows_dimension.valid_elements)
-        subtotal_fills = tuple(st.fill for st in self._rows_dimension.subtotals)
-        return tuple(
-            # ---Subtotals have negative sequential indexes (-1, -2, ..., -m)---
-            # ---To index them properly, we need to convert those indexes to---
-            # ---zero based positive indexes (0, 1, ... m - 1) i.e. -idx - 1---
-            (
-                element_fills[idx]
-                if idx > -1
-                else subtotal_fills[idx + len(subtotal_fills)]
-            )
-            for idx in self._row_order_signed_indexes
-        )
-
-    @lazyproperty
-    def scale_mean(self):
-        """Optional float mean of row numeric-values (scale).
-
-        This value is `None` when no row-elements have numeric-value. The numeric value
-        (aka. "scaled-count") for a row is its count multiplied by the numeric-value of
-        its element. For example, if 100 women responded "Very Likely" and the
-        numeric-value of the "Very Likely" response (element) was 4, then the
-        scaled-count for that row would be 400. The scale mean is the average of those
-        scale values over the total count of responses. The count of a row lacking a
-        numeric value does not contribute to either the numerator or denominator of that
-        computation.
-        """
-        return self._measures.scaled_counts.scale_mean
-
-    @lazyproperty
-    def scale_median(self):
-        """Optional float/int median of scaled weighted-counts.
-
-        This value is `None` when no rows have a numeric-value assigned. The median is
-        equal to a count of 1 multiplied by one of the row numeric-values, so the value
-        is equal to one of the assigned numeric-values (and retains its int/float type).
-        """
-        return self._measures.scaled_counts.scale_median
-
-    @lazyproperty
-    def scale_stddev(self):
-        """Optional np.float64 standard-deviation of scaled weighted-counts.
-
-        This value is `None` when no rows have a numeric-value assigned. The value has
-        the same units as the assigned numeric values and indicates the dispersion of
-        the scaled-count distribution from its mean (scale-mean).
-        """
-        return self._measures.scaled_counts.scale_stddev
-
-    @lazyproperty
-    def scale_stderr(self):
-        """Optional np.float64 standard-error of scaled weighted counts.
-
-        This value is `None` when no rows have a numeric-value assigned.
-        """
-        return self._measures.scaled_counts.scale_stderr
-
-    @lazyproperty
-    def share_sum(self):
-        """1D np.float64 ndarray of share of sum for each row.
-
-        Raises ValueError when the cube-result does not include a sum cube-measure.
-        """
-        return self._assemble_vector(self._measures.share_sum.blocks)
-
-    @lazyproperty
-    def smoothed_means(self):
-        """1D np.float64 ndarray of smoothed mean for each row.
-
-        Raises ValueError when the cube-result does not include a means cube-measure.
-        """
-        return self._assemble_vector(self._measures.smoothed_means.blocks)
-
-    @lazyproperty
-    def stddev(self):
-        """1D np.float64 ndarray of stddev for each row.
-
-        Raises ValueError when the cube-result does not include a stddev cube-measure.
-        """
-        return self._assemble_vector(self._measures.stddev.blocks)
-
-    @lazyproperty
-    def sums(self):
-        """1D np.float64 ndarray of sum for each row.
-
-        Raises ValueError when the cube-result does not include a sum cube-measure.
-        """
-        return self._assemble_vector(self._measures.sums.blocks)
-
-    @lazyproperty
-    def table_base_range(self):
-        """[min, max] np.float64 ndarray range of (total) unweighted-N for this stripe.
-
-        A non-MR stripe will have a single base, represented by min and max being the
-        same value. Each row of an MR stripe has a distinct base, which is reduced to a
-        range in that case.
-        """
-        return self._measures.unweighted_bases.table_base_range
-
-    @lazyproperty
-    def table_margin_range(self):
-        """[min, max] np.float64 ndarray range of (total) unweighted-N for this stripe.
-
-        A non-MR stripe will have a single margin, represented by min and max being the
-        same value. Each row of an MR stripe has a distinct margin, which is reduced to
-        a min/max range in that case.
-        """
-        return self._measures.weighted_bases.table_margin_range
-
-    @lazyproperty
-    def table_proportion_stddevs(self):
-        """1D np.float64 ndarray of table-proportion std-deviation for each row."""
-        return self._assemble_vector(self._measures.table_proportion_stddevs.blocks)
-
-    @lazyproperty
-    def table_proportion_stderrs(self):
-        """1D np.float64 ndarray of table-proportion std-error for each row."""
-        return self._assemble_vector(self._measures.table_proportion_stderrs.blocks)
-
-    @lazyproperty
-    def table_proportions(self):
-        """1D np.float64 ndarray of fraction of weighted-N contributed by each row."""
-        return self._assemble_vector(self._measures.table_proportions.blocks)
-
-    @lazyproperty
-    def unweighted_bases(self):
-        """1D np.float64 ndarray of unweighted table-proportion denominator per row."""
-        return self._assemble_vector(self._measures.unweighted_bases.blocks)
-
-    @lazyproperty
-    def unweighted_counts(self):
-        """1D np.float64 ndarray of unweighted count for each row of stripe."""
-        return self._assemble_vector(self._measures.unweighted_counts.blocks)
-
-    @lazyproperty
-    def weighted_bases(self):
-        """1D np.float64 ndarray of weighted table-proportion denominator per row."""
-        return self._assemble_vector(self._measures.weighted_bases.blocks)
-
-    @lazyproperty
-    def weighted_counts(self):
-        """1D np.float64 ndarray of weighted count for each row of stripe."""
-        return self._assemble_vector(self._measures.weighted_counts.blocks)
-
-    def _assemble_vector(self, blocks):
-        """Return 1D ndarray of base_vector with inserted subtotals, in order.
-
-        `blocks` is a pair of two 1D arrays, first the base-values and then the subtotal
-        values of the stripe vector. The returned array is sequenced in the computed
-        row order including possibly removing hidden or pruned values.
-        """
-        return np.concatenate(blocks)[self._row_order_signed_indexes]
-
-    @lazyproperty
-    def _measures(self):
-        """StripeMeasures collection object for this stripe."""
-        return StripeMeasures(
-            self._cube, self._rows_dimension, self._ca_as_0th, self._slice_idx
-        )
-
-    @lazyproperty
-    def _row_order_bogus_ids(self):
-        """Row order with bogus ids."""
-        return np.array(
-            _BaseOrderHelper.display_order(
-                self._rows_dimension, self._measures, format=ORDER_FORMAT.BOGUS_IDS
-            )
-        )
-
-    @lazyproperty
-    def _row_order_signed_indexes(self):
-        """Row order idx with signed idxs."""
-        return np.array(
-            _BaseOrderHelper.display_order(
-                self._rows_dimension, self._measures, format=ORDER_FORMAT.SIGNED_INDEXES
-            ),
-            dtype=int,
-        )
 
 
 # === ORDER HELPERS ===
